requirements.txt
dnspython==1.16.0
PyNaCl==1.3.0
async-timeout
discord.py
requests
git+https://github.com/mathsman5133/coc.py
gspread
oauth2client
pprint
asyncpg
#pandas